+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: sabrugie <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/08 11:14:06 by sabrugie          #+#    #+#             */
-/*   Updated: 2019/10/10 10:26:15 by sabrugie         ###   ########.fr       */
+/*   Updated: 2019/10/16 16:09:48 by sabrugie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strdup(const char *s1)
 {
-	unsigned int	i;
-	char			*s2;
+	size_t		i;
+	char		*s2;
 
 	i = 0;
 	while (s1[i])
